+++ src/pages/sujinAhn/Main/Contents/Components/Comments.js
@@ -1,0 +1,24 @@
+import { useState } from "react";
+
+const Comments = ({ list, removeComment }) => {
+  const [active, setActive] = useState(true);
+
+  const activeButton = () => {
+    setActive(!active);
+  };
+
+  return (
+    <li>
+      <span className="id">{list.name}</span>
+      <p className="comment-text">{list.text}</p>
+      <button className="btn-like" onClick={activeButton}>
+        <i className={active ? "fa-regular fa-heart" : "fa-solid fa-heart"} />
+      </button>
+      <button className="btn-delete" onClick={() => removeComment(list.id)}>
+        <i className="fa-solid fa-trash-can" />
+      </button>
+    </li>
+  );
+};
+
+export default Comments;
