+++ src/pages/sujinAhn/Main/Contents/Contents.js
@@ -1,0 +1,13 @@
+import Feeds from "./Components/Feeds";
+import Etc from "./Components/Etc";
+
+const Contents = () => {
+  return (
+    <main className="container">
+      <Feeds />
+      <Etc />
+    </main>
+  );
+};
+
+export default Contents;
